moved block to separate plugin
