+++ src/main.rs
@@ -1,23 +1,15 @@
 use bevy::prelude::*;
 
 use bevy::{
-    core_pipeline::CorePipelinePlugin,
-    diagnostic::DiagnosticsPlugin,
-    input::InputPlugin,
-    log::LogPlugin,
-    pbr::PbrPlugin,
-    render::{
-        render_resource::{Extent3d, TextureDimension, TextureFormat},
-        RenderPlugin,
-    },
-    scene::ScenePlugin,
-    sprite::SpritePlugin,
-    text::TextPlugin,
-    time::TimePlugin,
-    ui::UiPlugin,
-    window::WindowPlugin,
-    winit::WinitPlugin,
+    core_pipeline::CorePipelinePlugin, diagnostic::DiagnosticsPlugin, input::InputPlugin,
+    log::LogPlugin, pbr::PbrPlugin, render::RenderPlugin, scene::ScenePlugin, sprite::SpritePlugin,
+    text::TextPlugin, time::TimePlugin, ui::UiPlugin, window::WindowPlugin, winit::WinitPlugin,
 };
+
+mod block;
+mod texture;
+
+use block::BlockPlugin;
 
 fn main() {
     App::new()
@@ -46,35 +38,12 @@ fn main() {
         .add_plugin(UiPlugin::default())
         .add_plugin(TextPlugin::default())
         .add_plugin(WinitPlugin::default())
+        .add_plugin(BlockPlugin)
         .add_startup_system(setup)
         .run();
 }
 
-#[derive(Component)]
-struct Shape;
-
-fn setup(
-    mut commands: Commands,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut images: ResMut<Assets<Image>>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
-) {
-    let debug_material = materials.add(StandardMaterial {
-        base_color_texture: Some(images.add(uv_debug_texture())),
-        ..default()
-    });
-
-    let cube = meshes.add(shape::Cube::default().into());
-
-    commands.spawn((
-        PbrBundle {
-            mesh: cube,
-            material: debug_material,
-            ..default()
-        },
-        Shape,
-    ));
-
+fn setup(mut commands: Commands) {
     commands.spawn(PointLightBundle {
         point_light: PointLight {
             intensity: 9000.0,
@@ -91,32 +60,4 @@ fn setup(
             .looking_at(Vec3::new(0.0, 0.0, 0.0), Vec3::Y),
         ..default()
     });
-}
-
-/// Creates a colorful test pattern
-fn uv_debug_texture() -> Image {
-    const TEXTURE_SIZE: usize = 8;
-
-    let mut palette: [u8; 32] = [
-        255, 102, 159, 255, 255, 159, 102, 255, 236, 255, 102, 255, 121, 255, 102, 255, 102, 255,
-        198, 255, 102, 198, 255, 255, 121, 102, 255, 255, 236, 102, 255, 255,
-    ];
-
-    let mut texture_data = [0; TEXTURE_SIZE * TEXTURE_SIZE * 4];
-    for y in 0..TEXTURE_SIZE {
-        let offset = TEXTURE_SIZE * y * 4;
-        texture_data[offset..(offset + TEXTURE_SIZE * 4)].copy_from_slice(&palette);
-        palette.rotate_right(4);
-    }
-
-    Image::new_fill(
-        Extent3d {
-            width: TEXTURE_SIZE as u32,
-            height: TEXTURE_SIZE as u32,
-            depth_or_array_layers: 1,
-        },
-        TextureDimension::D2,
-        &texture_data,
-        TextureFormat::Rgba8UnormSrgb,
-    )
 }
